Fix avg_d

+++ realworld_benchmark/main_SBMs_node_classification.py
@@ -429,7 +429,7 @@ def main():
     net_params['n_classes'] = torch.unique(dataset.train[0][1],dim=0).size(0)
 
 
-    D = torch.cat([dataset.train[i][0].number_of_nodes() for i in range(len(dataset.train))])
+    D = torch.cat([torch.FloatTensor(dataset.train[i][0].number_of_nodes()) for i in range(len(dataset.train))])
     net_params['avg_d'] = dict(lin=torch.mean(D),
                                exp=torch.mean(torch.exp(torch.div(1, D)) - 1),
                                log=torch.mean(torch.log(D + 1)))
